backend/auth/requirements-auth.txt
fastapi>=0.68.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
sqlalchemy>=1.4.0
alembic>=1.7.0
